various improvements
- Make sure that stack-traversal respects the upstream if avaialbla
- Handle change-ids when rewriting references
- Factor code into module files to create space for growth
- also rewrite refs undere `refs/gitbutler/`
